refactor(core): split types.ts into src/core/interfaces/ folder
