+++ src/core/interfaces/Client.ts
@@ -1,0 +1,5 @@
+import type { SendRequestOptions } from './SendRequestOptions';
+
+export interface Client {
+  sendRequest<T>(options: SendRequestOptions<T>): Promise<T>;
+}
