+++ src/core/interfaces/ClientConfig.ts
@@ -1,14 +1,3 @@
-import type { ZodType } from 'zod';
-
-export interface SendRequestOptions<T = unknown> {
-  url: string;
-  method?: string;
-  headers?: Record<string, string>;
-  body?: unknown;
-  searchParams?: Record<string, unknown>;
-  schema?: ZodType<T>;
-}
-
 export interface ClientConfig {
   host?: string;
   apiKey: string;
@@ -20,8 +9,4 @@ export interface ClientConfig {
    * skip parsing overhead on large responses. Defaults to `false`.
    */
   skipParsing?: boolean;
-}
-
-export interface Client {
-  sendRequest<T>(options: SendRequestOptions<T>): Promise<T>;
 }
