+++ src/core/interfaces/index.ts
@@ -1,0 +1,5 @@
+export type { Client } from './Client';
+
+export type { ClientConfig } from './ClientConfig';
+
+export type { SendRequestOptions } from './SendRequestOptions';
